+++ tests/testthat/test-bed-clumping-SVD.R
@@ -19,9 +19,9 @@ obj.bed <- bed(system.file("extdata", "example.bed", package = "bigsnpr"))
 expect_error(bed_clumping(G))
 ind.keep2 <- bed_clumping(obj.bed)
 expect_identical(ind.keep2, ind.keep)
-expect_error(bed_randomSVD(G, ind.col = ind.keep),
-             "'obj.bed' is not of class 'bed'.")
-obj.svd2 <- bed_randomSVD(obj.bed, ind.col = ind.keep)
+expect_error(big_randomSVD(obj.bed, ind.col = ind.keep),
+             "'X' is not of class 'FBM'.")
+obj.svd2 <- big_randomSVD(obj.bed, bed_scaleBinom, ind.col = ind.keep)
 expect_equal(colMeans(obj.svd2$u), rep(0, 10))
 expect_equal(obj.svd2, obj.svd)
 
@@ -31,7 +31,7 @@ obj.bed2 <- bed(snp_writeBed(bigSNP, tempfile(fileext = ".bed")))
 ind.keep3 <- bed_clumping(obj.bed2)
 expect_gt(length(intersect(ind.keep3, ind.keep2)) /
             length(union(ind.keep3, ind.keep2)), 0.95)
-obj.svd3 <- bed_randomSVD(obj.bed2, ind.col = ind.keep2)
+obj.svd3 <- big_randomSVD(obj.bed2, bed_scaleBinom, ind.col = ind.keep2)
 expect_gt(mean(sqrt(colSums(cor(obj.svd3$u, obj.svd2$u)^2))), 0.9)
 expect_equal(obj.svd3$d, obj.svd2$d, tolerance = 0.1)
 expect_true(all(obj.svd3$d < obj.svd2$d))
@@ -51,7 +51,8 @@ not_cran <- identical(Sys.getenv("BIGSNPR_CRAN"), "false")
 NCORES <- `if`(not_cran, 2, 1)
 
 expect_identical(bed_clumping(obj.bed2, ncores = NCORES), ind.keep3)
-expect_equal(bed_randomSVD(obj.bed2, ind.col = ind.keep2, ncores = NCORES),
+expect_equal(big_randomSVD(obj.bed2, bed_scaleBinom,
+                           ind.col = ind.keep2, ncores = NCORES),
              obj.svd3)
 
 ################################################################################